+++ jbt/src/main/java/org/hibernate/tool/orm/jbt/internal/util/DummyMetadataBuildingContext.java
@@ -11,7 +11,6 @@ import org.hibernate.boot.spi.InFlightMetadataCollector;
 import org.hibernate.boot.spi.MetadataBuildingContext;
 import org.hibernate.boot.spi.MetadataBuildingOptions;
 import org.hibernate.cfg.AvailableSettings;
-import org.hibernate.tool.orm.jbt.util.MockDialect;
 
 public class DummyMetadataBuildingContext {
