+++ jbt/src/test/java/org/hibernate/tool/orm/jbt/api/wrp/SchemaExportWrapperTest.java
@@ -15,7 +15,7 @@ import org.hibernate.tool.hbm2ddl.SchemaExport;
 import org.hibernate.tool.orm.jbt.internal.factory.ConfigurationWrapperFactory;
 import org.hibernate.tool.orm.jbt.internal.factory.SchemaExportWrapperFactory;
 import org.hibernate.tool.orm.jbt.internal.util.MockConnectionProvider;
-import org.hibernate.tool.orm.jbt.util.MockDialect;
+import org.hibernate.tool.orm.jbt.internal.util.MockDialect;
 import org.hibernate.tool.schema.TargetType;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
