+++ jbt/src/test/java/org/hibernate/tool/orm/jbt/internal/util/ConfigurationMetadataDescriptorTest.java
@@ -10,7 +10,6 @@ import java.util.Properties;
 import org.hibernate.boot.Metadata;
 import org.hibernate.cfg.AvailableSettings;
 import org.hibernate.cfg.Configuration;
-import org.hibernate.tool.orm.jbt.util.MockDialect;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
