+++ jbt/src/test/java/org/hibernate/tool/orm/jbt/internal/util/MockDialectTest.java
@@ -1,8 +1,9 @@
-package org.hibernate.tool.orm.jbt.util;
+package org.hibernate.tool.orm.jbt.internal.util;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 
 import org.hibernate.dialect.DatabaseVersion;
+import org.hibernate.tool.orm.jbt.internal.util.MockDialect;
 import org.junit.jupiter.api.Test;
 
 public class MockDialectTest {
